% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eurovisionsongcontest-package.R
\docType{package}
\name{eurovisionsongcontest}
\alias{eurovisionsongcontest}
\alias{eurovisionsongcontest-package}
\title{Scoreboards of the Eurovision Song Contest}
\description{
Scoreboards of the Eurovision Song Contest
}

