% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eurovisionsongcontest-package.R
\docType{package}
\name{eurovisionsongcontest}
\alias{eurovisionsongcontest}
\alias{eurovisionsongcontest-package}
\title{Scoreboards of the Eurovision Song Contest}
\format{\preformatted{Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	42824 obs. of  5 variables:
 $ Participant: chr  "The Netherlands" "Belgium" "Germany" "France" ...
 $ Event      : chr  "Eurovision Song Contest 1956" "Eurovision Song Contest 1956" "Eurovision Song Contest 1956" "Eurovision Song Contest 1956" ...
 $ Date       : Date, format: "1956-05-24" "1956-05-24" ...
 $ key        : chr  "Place" "Place" "Place" "Place" ...
 $ value      : num  2 2 2 2 2 2 1 0 3 1 ...
}}
\usage{
eurovisionsongcontest
}
\description{
Scoreboards of the Eurovision Song Contest scraped from the official website.
}
\keyword{datasets}

