% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape}
\alias{scrape}
\title{Scrape an Event}
\usage{
scrape(event_url, sleep = 0)
}
\arguments{
\item{event_url}{The event URL}

\item{sleep}{Amount of seconds to sleep}
}
\value{
A data frame
}
\description{
Scrape an Event
}
\examples{
\dontrun{
scrape("http://www.eurovision.tv/page/history/by-year/contest?event=2083")
}
}

